backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-jose
passlib[bcrypt]
motor
exa_py
